EDM-1882: internal/tpm: refactor
